requirements.txt
dhanhq==1.2.2
python-dotenv==1.0.0
pandas==2.1.3
pandas-ta==0.3.14b0
click>=8.1.7
